Add test to check multiple service deployment with same directory
The directory just above the actual docker-compose-file serves the
purpose of the `context` for docker-compose. Thus, when its time to down
the service, it downs all the services IT considers to be in the same
project/context. This means:

project_1
    server
        docker-compose.yml
project_2
    server
        docker-compose.yml

in above case, for docker-compose project_1 and project_2 will be part
of the same context. This test creates similar condition